test that wholesale results have been written to delta table
