+++ source/databricks/calculation-engine/tests/integration/calculator_job/test_calculator_job_balance_fixing_results.py
@@ -106,61 +106,6 @@ ALL_ENERGY_RESULT_TYPES = {
     ),
 }
 
-WHOLESALE_FIXING_ENERGY_RESULT_TYPES = {
-    (
-        TimeSeriesType.NET_EXCHANGE_PER_GA.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.PRODUCTION.value,
-        AggregationLevel.ES_PER_GA.value,
-    ),
-    (
-        TimeSeriesType.PRODUCTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.NON_PROFILED_CONSUMPTION.value,
-        AggregationLevel.ES_PER_GA.value,
-    ),
-    (
-        TimeSeriesType.NON_PROFILED_CONSUMPTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.FLEX_CONSUMPTION.value,
-        AggregationLevel.ES_PER_GA.value,
-    ),
-    (
-        TimeSeriesType.FLEX_CONSUMPTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.GRID_LOSS.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.POSITIVE_GRID_LOSS.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.NEGATIVE_GRID_LOSS.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.TOTAL_CONSUMPTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.TEMP_FLEX_CONSUMPTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-    (
-        TimeSeriesType.TEMP_PRODUCTION.value,
-        AggregationLevel.TOTAL_GA.value,
-    ),
-}
-
 
 @pytest.mark.parametrize(
     "time_series_type, aggregation_level", ALL_ENERGY_RESULT_TYPES,
@@ -197,40 +142,3 @@ def test__balance_fixing_result__has_expected_number_of_result_types(
 
     # Assert: The result is created if there are rows
     assert actual_result_type_count == len(ALL_ENERGY_RESULT_TYPES)
-
-
-@pytest.mark.parametrize(
-    "time_series_type, aggregation_level", WHOLESALE_FIXING_ENERGY_RESULT_TYPES,
-)
-def test__wholesale_result__is_created(
-    wholesale_fixing_results_df: DataFrame,
-    time_series_type: str,
-    aggregation_level: str,
-) -> None:
-    # Arrange
-    result_df = (
-        wholesale_fixing_results_df.where(F.col(EnergyResultColumnNames.batch_id) == C.executed_wholesale_batch_id)
-        .where(F.col(EnergyResultColumnNames.time_series_type) == time_series_type)
-        .where(F.col(EnergyResultColumnNames.aggregation_level) == aggregation_level)
-    )
-
-    # Act: Calculator job is executed just once per session. See the fixtures `results_df` and `executed_wholesale_fixing`
-
-    # Assert: The result is created if there are rows
-    assert result_df.count() > 0
-
-
-def test__wholesale_result__has_expected_number_of_energy_result_types(
-    wholesale_fixing_results_df: DataFrame,
-) -> None:
-    # Arrange
-    actual_result_type_count = (
-        wholesale_fixing_results_df.where(F.col(EnergyResultColumnNames.batch_id) == C.executed_wholesale_batch_id)
-        .where(F.col(EnergyResultColumnNames.batch_id) == C.executed_wholesale_batch_id)
-        .select(EnergyResultColumnNames.time_series_type, EnergyResultColumnNames.aggregation_level).distinct().count()
-    )
-
-    # Act: Calculator job is executed just once per session. See the fixtures `results_df` and `executed_wholesale_fixing`
-
-    # Assert: The result is created if there are rows
-    assert actual_result_type_count == len(WHOLESALE_FIXING_ENERGY_RESULT_TYPES)
